restructure project files
- create constants file for environment variables and other constants
- create events file for event registration
- create commands file for command registration
- update command parsing to use command registration
